# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TrendPoint {
  label: string;
  score: number;
}

export type InsightSeverity = typeof InsightSeverity[keyof typeof InsightSeverity];


export const InsightSeverity = {
  healthy: 'healthy',
  attention: 'attention',
  high: 'high',
  critical: 'critical',
  neutral: 'neutral',
} as const;

export interface Insight {
  title: string;
  context: string;
  evidence: string;
  recommendation: string;
  severity: InsightSeverity;
}

export interface Dashboard {
  companyName: string;
  cycleLabel: string;
  overallScore: number;
  responseRate: number;
  employeesEvaluated: number;
  sectorsEvaluated: number;
  criticalRisks: number;
  pendingActions: number;
  trend: TrendPoint[];
  highlights: Insight[];
}

export type RiskTrend = typeof RiskTrend[keyof typeof RiskTrend];


export const RiskTrend = {
  up: 'up',
  down: 'down',
  stable: 'stable',
} as const;

export type RiskStatus = typeof RiskStatus[keyof typeof RiskStatus];


export const RiskStatus = {
  healthy: 'healthy',
  attention: 'attention',
  high: 'high',
  critical: 'critical',
  neutral: 'neutral',
} as const;

export interface Risk {
  id: number;
  factor: string;
  sector: string;
  /** Scale from 1 to 5 */
  probability: number;
  /** Scale from 1 to 5 */
  severity: number;
  exposure: string;
  trend: RiskTrend;
  status: RiskStatus;
  recommendation: string;
}

export type ActionPlanStatus = typeof ActionPlanStatus[keyof typeof ActionPlanStatus];


export const ActionPlanStatus = {
  planned: 'planned',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type ActionPlanPriority = typeof ActionPlanPriority[keyof typeof ActionPlanPriority];


export const ActionPlanPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface ActionPlan {
  id: number;
  title: string;
  owner: string;
  dueDate: string;
  status: ActionPlanStatus;
  priority: ActionPlanPriority;
  source: string;
}

export type ActionPlanInputPriority = typeof ActionPlanInputPriority[keyof typeof ActionPlanInputPriority];


export const ActionPlanInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface ActionPlanInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  owner: string;
  /** @minLength 1 */
  dueDate: string;
  priority: ActionPlanInputPriority;
  /** @minLength 1 */
  source: string;
}

export interface AssessmentQuestion {
  id: number;
  text: string;
  options: string[];
}

export interface Assessment {
  id: number;
  title: string;
  description: string;
  privacyNote: string;
  questions: AssessmentQuestion[];
}

export type AssessmentResponseInputAnswers = {[key: string]: string};

export interface AssessmentResponseInput {
  assessmentId: number;
  answers: AssessmentResponseInputAnswers;
}

export interface AssessmentResponse {
  accepted: boolean;
  message: string;
}