# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /action-plans:
    get:
      operationId: getActionPlans
      tags: [actions]
      summary: List action plans
      responses:
        "200":
          description: Action plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActionPlan"
    post:
      operationId: createActionPlan
      tags: [actions]
      summary: Create an action plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActionPlanInput"
      responses:
        "201":
          description: Created action plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionPlan"
...
  /assessment/responses:
    post:
      operationId: submitAssessmentResponse
      tags: [assessment]
      summary: Submit an anonymous assessment response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentResponseInput"
      responses:
        "201":
          description: Response accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResponse"
components:
  schemas:
...
    ActionPlan:
      type: object
      required: [id, title, owner, dueDate, status, priority, source]
      properties:
        id: { type: number }
        title: { type: string }
        owner: { type: string }
        dueDate: { type: string }
        status: { type: string, enum: [planned, in_progress, done] }
        priority: { type: string, enum: [low, medium, high] }
        source: { type: string }
    ActionPlanInput:
      type: object
      required: [title, owner, dueDate, priority, source]
      properties:
        title: { type: string, minLength: 1 }
        owner: { type: string, minLength: 1 }
        dueDate: { type: string, minLength: 1 }
        priority: { type: string, enum: [low, medium, high] }
        source: { type: string, minLength: 1 }
...
    AssessmentResponseInput:
      type: object
      required: [assessmentId, answers]
      properties:
        assessmentId: { type: number }
        answers:
          type: object
          additionalProperties: { type: string }
    AssessmentResponse:
      type: object
      required: [accepted, message]
      properties:
        accepted: { type: boolean }
        message: { type: string }