# lib/api-zod/src/generated/types/actionPlan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionPlanPriority } from './actionPlanPriority';
import type { ActionPlanStatus } from './actionPlanStatus';

export interface ActionPlan {
  id: number;
  title: string;
  owner: string;
  dueDate: string;
  status: ActionPlanStatus;
  priority: ActionPlanPriority;
  source: string;
}