# lib/api-zod/src/generated/types/actionPlanInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActionPlanInputPriority } from './actionPlanInputPriority';

export interface ActionPlanInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  owner: string;
  /** @minLength 1 */
  dueDate: string;
  priority: ActionPlanInputPriority;
  /** @minLength 1 */
  source: string;
}