# lib/api-zod/src/generated/types/assessmentResponseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentResponseInputAnswers } from './assessmentResponseInputAnswers';

export interface AssessmentResponseInput {
  assessmentId: number;
  answers: AssessmentResponseInputAnswers;
}